[runtime] Enhancement: Scheduler Cleanup
